Organized directory structure and lint
