+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "editor.formatOnSave": true,
+  "eslint.autoFixOnSave": true
+}
