+++ src/assets/css/style.css
@@ -35,7 +35,6 @@ body {
   color: #ffffff;
   border: 1px solid #999999;
   box-shadow: 1px 1px 1px #aaaaaa;
-
 }
 
 .left {
