+++ src/components/Button.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style.css';
+import '../assets/css/style.css';
 
 function Button({ style, onClick, text }) {
   return (
