+++ src/components/Chatting.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style.css';
+import '../assets/css/style.css';
 
 function Chatting({ style, text }) {
   return <p className={style}>{text}</p>;
