+++ src/components/ChattingBox.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import './style.css';
 import { Container } from 'react-bootstrap';
+import '../assets/css/style.css';
 import Chatting from './Chatting';
 
 const chattings = [
@@ -41,9 +41,9 @@ const chattings = [
 function ChattingBox() {
   return (
     <Container className="chatting_box">
-      {chattings.map((chatting, i) => {
-        return <Chatting text={chatting.text} style={chatting.style} key={i} />;
-      })}
+      {chattings.map((chatting, i) => (
+        <Chatting text={chatting.text} style={chatting.style} key={i} />
+      ))}
     </Container>
   );
 }
