+++ src/components/Home.js
@@ -1,34 +1,34 @@
-import React, { Component } from 'react';
-import Logo from './Logo';
+import React from 'react';
+import { Container, Row, Col } from 'react-bootstrap';
+import Logo from '../assets/img/logo.jpg';
 import Title from './Title';
 import Login from './Login';
-import './style.css';
+import '../assets/css/style.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { Container, Row, Col } from 'react-bootstrap';
 
 function Home() {
   return (
     <Container>
       <Row>
-        <Col></Col>
+        <Col />
         <Col>
           <Logo />
         </Col>
-        <Col></Col>
+        <Col />
       </Row>
       <Row>
-        <Col></Col>
+        <Col />
         <Col>
           <Title />
         </Col>
-        <Col></Col>
+        <Col />
       </Row>
       <Row>
-        <Col></Col>
+        <Col />
         <Col>
           <Login />
         </Col>
-        <Col></Col>
+        <Col />
       </Row>
     </Container>
   );
