+++ src/components/InputForm.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import './style.css';
 import { Container, Row, Col } from 'react-bootstrap';
 import InputText from './InputText';
 import Button from './Button';
+import '../assets/css/style.css';
 
 function InputForm() {
   return (
