+++ src/components/InputText.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import './style.css';
 import { Form } from 'react-bootstrap';
+import '../assets/css/style.css';
 
 function InputText() {
   return (
