+++ src/components/Logo.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './style.css';
 import { Image } from 'react-bootstrap';
-import Logo from './logo.jpg';
+import Logo from '../assets/img/logo.jpg';
+import '../assets/css/style.css';
 
 function Logo() {
   return <Image src={Logo} alt="logo" />;
