+++ src/components/Users.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import User from './User';
-import './style.css';
 import { Container } from 'react-bootstrap';
+import User from './User';
+import '../assets/css/style.css';
 
 const users = [
   {
@@ -75,9 +75,15 @@ const users = [
 function Users() {
   return (
     <Container className="users">
-      {users.map(({ userName, icon, flag, language }, index) => {
-        return <User userName={userName} icon={icon} flag={flag} language={language} key={index} />;
-      })}
+      {users.map(({ userName, icon, flag, language }, index) => (
+        <User
+          userName={userName}
+          icon={icon}
+          flag={flag}
+          language={language}
+          key={index}
+        />
+      ))}
     </Container>
   );
 }
